move string/struct extensions into extensions_for namespace
