+++ lib/trax/model/extensions_for/string.rb
@@ -1,0 +1,13 @@
+module Trax
+  module Model
+    module ExtensionsFor
+      module String
+        extend ::ActiveSupport::Concern
+
+        def uuid
+          ::Trax::Model::UUID === self ? ::Trax::Model::UUID.new(self) : nil
+        end
+      end
+    end
+  end
+end
